requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Flask==2.2.3
flask-expects-json==1.7.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
psycopg2-binary==2.9.5
pyrsistent==0.19.3
requests==2.28.2
urllib3==1.26.14
Werkzeug==2.2.3
